Update MEM related documentation
This patch set updates the following aspects of the Midokura
Enterprise MidoNet (MEM) related documentation:

1) Point to correct version of external docs

Eucalyptus currently only supports MEM v1.9.

Change the external links to Midokura's documentation to point to
v1.9 instead of the latest version (which is v5.2 at the time of
writing).

2) Point to same set of external docs

Most of the external links point to Midokura's "RHEL 7 / Kilo (RDO)"
set of the documentation, just one link pointed to "RHEL 7 / Juno
(OSP)" instead. Unify this.

3) Update repository access information

To access the Midokura repository, access credentials are necessary.

Signing up for those should not be done via support.midokura.com,
which is an internal portal for technical support of existing
customers.

Add an email address to contact the Midokura sales department
instead.

Signed-off-by: Jan Hilberath <jan@midokura.com>